+++ djmoney/contrib/exchange/backends/__init__.py
@@ -1,2 +1,3 @@
 from .fixer import FixerBackend  # noqa
 from .openexchangerates import OpenExchangeRatesBackend  # noqa
+from .unirateapi import UniRateBackend  # noqa