+++ tests/settings.py
@@ -55,6 +55,7 @@ moneyed.add_currency(code="USDT", numeric="000", name="Tether")
 
 OPEN_EXCHANGE_RATES_APP_ID = "test"
 FIXER_ACCESS_KEY = "test"
+UNIRATE_ACCESS_KEY = "test"
 
 DEFAULT_CURRENCY = "USD"