+++ docs/changes.rst
@@ -2,6 +2,14 @@ Changelog
 =========
 
 
+`Unreleased`_
+-------------
+
+**Added**
+
+- ``UniRateBackend`` exchange-rate backend for `UniRateAPI <https://unirateapi.com>`_. Configure via ``EXCHANGE_BACKEND = "djmoney.contrib.exchange.backends.UniRateBackend"`` and set ``UNIRATE_ACCESS_KEY``.
+
+
 `3.6`_ - 2026-01-27
 -------------------